requirements.txt
flask
python-dotenv

langchain==0.1.20
langchain-community==0.0.38
langchain-core==0.1.52
langchain-openai==0.1.7
langchain-text-splitters==0.0.1

faiss-cpu
pypdf
sentence-transformers
tiktoken
openai>=1.10.0
